Added SSL cryto and HTTP
SSL Crypto probably won't work depending on the version of openssl. It
is just added as an example.

HTTP will send an http request to a host, if the response is successful
(200 OK or some other 2xx response), then it will print show the host's
http headers contained in the response.