Added OWASP ZAP pen testing
Use OWASP ZAP's spider to crawl through a webpage and generate a
list/map of available pages on a server. Requires configuration in
config.ini (moved in the same directory as the source code). For some
reason, after including this change in the project, I needed to run this
as a super user.